+++ example/freertos_feature/task_notify/main.c
@@ -56,7 +56,7 @@ int main(void)
     taskENTER_CRITICAL(); /*进入临界区*/
     ret = xTaskCreate((TaskFunction_t)NotifyExampleTaskEntry,  /* 任务入口函数 */
                           (const char *)"NotifyExampleTaskEntry",/* 任务名字 */
-                          (uint16_t)4096,  /* 任务栈大小 */
+                          4096,  /* 任务栈大小 */
                           NULL,/* 任务入口函数参数 */
                           (UBaseType_t)8,  /* 任务的优先级 */
                           NULL);
