+++ example/peripheral/i2c/src/i2c_ms_example.c
@@ -485,7 +485,7 @@ BaseType_t FFreeRTOSI2cLoopbackCreate(void)
 
     xReturn = xTaskCreate((TaskFunction_t)FFreeRTOSI2cLoopbackTask,  /* 任务入口函数 */
                           (const char *)"FFreeRTOSI2cLoopbackTask",/* 任务名字 */
-                          (uint16_t)1024,  /* 任务栈大小 */
+                          1024,  /* 任务栈大小 */
                           (void *)NULL,/* 任务入口函数参数 */
                           (UBaseType_t)I2C_MS_TEST_TASK_PRIORITY,  /* 任务的优先级 */
                           NULL); /* 任务控制 */
