+++ example/system/amp/openamp/driver_core/main.c
@@ -58,7 +58,7 @@ int main(void)
     taskENTER_CRITICAL(); /* no schedule when create task */
     ret = xTaskCreate((TaskFunction_t )OpenampExampleTaskEntry, /* 任务入口函数 */
                         (const char* )"OpenampExampleTaskEntry",/* 任务名字 */
-                        (uint16_t )(4096*2), /* 任务栈大小 */
+                        (4096*2), /* 任务栈大小 */
                         (void* )NULL,/* 任务入口函数参数 */
                         (UBaseType_t )4, /* 任务的优先级 */
                         NULL); /* 任务控制块指针 */
