+++ example/system/nested_interrupt/src/nested_interrupt.c
@@ -191,7 +191,7 @@ BaseType_t FFreeRTOSNestedIntrTaskCreate(void)
 
     xReturn = xTaskCreate((TaskFunction_t)NestedIntrTask, /* 任务入口函数 */
                           (const char *)"NestedIntrTask", /* 任务名字 */
-                          (uint16_t)TASK_STACK_SIZE,                            /* 任务栈大小 */
+                          TASK_STACK_SIZE,                            /* 任务栈大小 */
                           NULL,                                      /* 任务入口函数参数 */
                           (UBaseType_t)NESTED_INTR_TEST_TASK_PRIORITY,     /* 任务的优先级 */
                           NULL);
